third_party/blink/web_tests/wpt_internal/webgpu/canvas_webgpu_transfer/transferToGPUTexture-unbalanced-access.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./canvas-test.js"></script>
<script src="./webgpu-helpers.js"></script>

<h1>Canvas2D WebGPU Transfer: transferToGPUTexture-unbalanced-access</h1>
<p>Unbalanced calls to transferToGPUTexture() will destroy the old WebGPU access
texture.</p>

<script>

runCanvasTestsInWorker({dependencies: ['./webgpu-helpers.js']});

/**
 * Unbalanced calls to transferToGPUTexture() will destroy the old WebGPU access
 * texture.
 */
canvasPromiseTest(async (canvas) => {
  const adapter = await getWebGPUAdapter();
  // Skip this test on Mac Swiftshader.
  if (isMacSwiftShader(adapter.info)) {
    return;
  }

  const device = await getWebGPUDevice(adapter);

  // Begin a WebGPU access session.
  const ctx = canvas.getContext('2d');
  const tex1 = ctx.transferToGPUTexture({device: device,
                                         usage: GPUTextureUsage.COPY_DST |
                                                GPUTextureUsage.COPY_SRC});

  // Start a second WebGPU access session, destroying the first texture.
  const tex2 = ctx.transferToGPUTexture({device: device,
                                         usage: GPUTextureUsage.COPY_DST |
                                                GPUTextureUsage.COPY_SRC});

  // Only the second texture should remain in an undestroyed state.
  assert_true(await isTextureDestroyed(device, tex1));
  assert_false(await isTextureDestroyed(device, tex2));
}, 'Unbalanced calls to transferToGPUTexture() will destroy the old WebGPU ' +
   'access texture.');

</script>
